+++ src/main/java/frogger/model/WetTurtle.java
@@ -7,7 +7,7 @@ public class WetTurtle extends Actor{
 	Image turtle2;
 	Image turtle3;
 	Image turtle4;
-	private int speed;
+	private double speed;
 	int i = 1;
 	boolean bool = true;
 	boolean sunk = false;
@@ -36,7 +36,7 @@ public class WetTurtle extends Actor{
 		if (getX() < -75 && speed<0)
 			setX(600);
 	}
-	public WetTurtle(int xpos, int ypos, int s, int w, int h) {
+	public WetTurtle(int xpos, int ypos, double s, int w, int h) {
 		 turtle1 = new Image("file:src/resources/TurtleAnimation1.png", w, h, true, true);
 		 turtle2 = new Image("file:src/resources/TurtleAnimation2Wet.png", w, h, true, true);
 		 turtle3 = new Image("file:src/resources/TurtleAnimation3Wet.png", w, h, true, true);
